backend/ml/requirements.txt
xgboost
scikit-learn
pandas
numpy
joblib
matplotlib
psycopg2-binary